impl/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "labtrust-portfolio"
version = "0.1.0"
description = "Thin-slice MAESTRO + Replay pipeline for the LabTrust paper portfolio."
requires-python = ">=3.10"
dependencies = ["jsonschema>=4.18.0", "pyyaml>=6.0", "python-dotenv>=1.0.0"]

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["labtrust_portfolio", "labtrust_portfolio.adapters"]
